Fixed error messages
The error messages now display in a slightly more attractive way, but there is also now a weird black area between the two buttons in the regular popup. The innerHTML of the systemLabels is now successfully saved to the `default` storage item and logged to console to make it easier to manipulate throughout development.